+++ Spidy/Assets/Scripts/LeftOrRight.cs
@@ -5,6 +5,12 @@ using System;
 
 public class LeftOrRight : MonoBehaviour
 {
+    void Start()
+    {
+        Spawn webSpawner = GetComponent<Spawn>();
+        webSpawner.OnRespawn += PutWebLeftOrRightRandomly; 
+    }
+    
     void PutWebLeftOrRightRandomly (object sender, Spawn.OnRespawnArgs e) 
     {
         e.webTransform.position += new Vector3 (LeftOrRightFromCenter(),0,0);
@@ -15,11 +21,4 @@ public class LeftOrRight : MonoBehaviour
         if ( UnityEngine.Random.Range(0, 2) != 0) return 1 ;
         return -1;
     }
-    void Start()
-    {
-        Spawn webSpawner = GetComponent<Spawn>();
-        webSpawner.OnRespawn += PutWebLeftOrRightRandomly; 
-    }
-
-    
 }